[FIX]carousel 컴포넌트화 및 css 수정
