+++ src/components/thema_lesson/ThemaLesson.module.scss
@@ -53,7 +53,13 @@
     justify-content: center;
     margin: 0px auto;
   }
-
+  .box {
+    width: 100%;
+    margin: 0px auto;
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+  }
   .wrap {
     margin: 5px;
     border-radius: 2%;
@@ -72,6 +78,13 @@
     display: flex;
     justify-content: center;
     margin: 0px auto;
+  }
+  .box {
+    width: 100%;
+    margin: 0px auto;
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
   }
 
   .wrap {
