+++ src/pages/slide/Banner.module.scss
@@ -1,16 +1,6 @@
 $slide-count: 3;
 $slide-width: 22em; //15em;;
 
-ul {
-  list-style-type: none;
-  padding: 0;
-}
-
-a {
-  text-decoration: none;
-  color: inherit;
-}
-
 .carouselComponent {
   width: 1000px;
   margin: 0px auto;
